requirements.txt
# Python Dependencies for ICD-10 Coding Project
# Install with: pip install -r requirements.txt

# Deep Learning
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Scientific Computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# NLP
nltk==3.8.1
spacy==3.7.2
transformers==4.35.2
gensim==4.4.0

# LLM APIs
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.0
litellm>=1.0.0

# Vector Search & RAG
faiss-cpu==1.7.4
llama-index==0.9.20
llama-index-embeddings-openai==0.1.1
llama-index-llms-openai==0.1.1
chromadb==0.4.18

# Data Processing
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0

# Evaluation
torchmetrics==1.2.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.2

# Utilities
requests==2.31.0
colorama==0.4.6
rich==13.6.0
python-dateutil==2.8.2
